Perform wrapObject recursively on JsonObject and JsonArray
Currently when we call wrapObject it applies to only the initial JsonObject/JsonArray.  This means that if there are nested JsonObject or JsonArray instances in the object graph they won't get converted.  This can lead to issues later on when Groovy code assumes it is dealing with only its native Json form (map/list).  This changes applies the conversion recursively.