Update templates to use wrapObject when converting JsonObject/JsonArray
We want to ensure that nested JsonObject and JsonArray instances are properly converted.  This is done by InternalHelper.wrapObject() so we want to use this method instead of directly extracting the map and/or list.

To help test this we added new methods to the codegen model and used those to add new test cases here in ApiTest.  This relies on https://github.com/vert-x3/vertx-codegen/pull/33.

Added a new test to EventBusTest to test sending and receiving nested Json objects.

Signed-off-by: Sean Fitts <smfitts@gmail.com>